Added seting up
