+++ go.mod
@@ -1,0 +1,8 @@
+module github.com/vetusbs/utils-api
+
+go 1.13
+
+require (
+	github.com/gorilla/mux v1.8.0
+	github.com/sirupsen/logrus v1.7.0
+)
