+++ heroku.yml
@@ -1,0 +1,6 @@
+build:
+  languages:
+    - go
+
+run:
+  web: api-utils
